% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangement_methods.R
\name{Arrangement}
\alias{Arrangement}
\title{Constructor function to arrange STLS}
\usage{
Arrangement(...)
}
\arguments{
\item{a}{nemaed vecrtor of colors / STLS}
}
\description{
Constructor function to arrange STLS
}
