% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_column.R
\name{case_vpartition}
\alias{case_vpartition}
\title{Seperate an STL into several colors / bodies by logical rules}
\usage{
case_vpartition(STL, ...)
}
\arguments{
\item{STL}{an STL object}

\item{...}{A series of two-sided formulas. As with dplyr's case_when function,
    the left-hand-side (LHS) determines which values match this case. The right-hand-side
    determines which STL the column should be separated into.

    formulas can have references to 'top', 'x', 'y', 'z' and 'bot.' 'z' is shorthand for 'top',
    which refers to the z value of the top surface at any point. 'x' and 'y' refer to the x and y coordinates
    at any point, and 'bot' refers to the z value of the bottom surface.}
}
\description{
Seperate an STL into several colors / bodies by logical rules
}
