% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{find_normal}
\alias{find_normal}
\title{A function to find the vector normal to a triangle
from three given vertexes}
\usage{
find_normal(vertex_1, vertex_2, vertex_3)
}
\arguments{
\item{vertex_1}{A triangle vertex}

\item{vertex_2}{A triangle vertex}

\item{vertex_3}{A triangle vertex}
}
\value{
A numeric vector
}
\description{
A function to find the vector normal to a triangle
from three given vertexes
}
