% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{normalize_vector}
\alias{normalize_vector}
\title{A function to find the unit vector for a given vector}
\usage{
normalize_vector(vect)
}
\arguments{
\item{...}{a numeric vector}
}
\value{
A numeric vector
}
\description{
A function to find the unit vector for a given vector
}
