% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writestl.R
\name{write_stl}
\alias{write_stl}
\title{A Fucntion to write a matrix to an STL File}
\usage{
write_stl(df, filename)
}
\arguments{
\item{df}{a matrix or matrix-like object}

\item{filename}{A the name of the file to save out. A srting.}
}
\value{
None
}
\description{
A Fucntion to write a matrix to an STL File
}
