% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{xprod}
\alias{xprod}
\title{A function to find the cross product of an arbitrary number of vectors}
\usage{
xprod(...)
}
\arguments{
\item{...}{a list of vectors}
}
\value{
A numeric vector
}
\description{
A function to find the cross product of an arbitrary number of vectors
}
